api/pyproject.toml
[project]
name = "api"
version = "1.0.0"
description = "FastAPI backend with React frontend"
requires-python = ">=3.11,<3.13"  # Constrain to Python 3.11-3.12 for stability
dependencies = [
    "fastapi>=0.104.0,<0.106.0",
    "uvicorn>=0.24.0,<0.26.0",
    "sqlalchemy>=2.0.23,<2.1.0",
    "aiosqlite>=0.19.0,<0.21.0",
    "python-jose[cryptography]>=3.3.0,<3.4.0",
    "passlib[bcrypt]>=1.7.4,<1.8.0",
    "python-multipart>=0.0.6,<0.1.0",
    "pydantic>=2.4.2,<2.6.0",
    "pydantic-settings>=2.0.0,<2.2.0",
    "bcrypt==4.0.1",  # Pin bcrypt version to resolve warning
    # Rate limiting - Updated to use redis.asyncio
    "fastapi-limiter>=0.1.5,<0.2.0",
    "redis>=5.0.0,<6.0.0",  # Replaces aioredis - asyncio support merged into redis-py
    "httpx>=0.24.0,<0.26.0",
    # ML dependencies - Updated for Python 3.12 compatibility
    "mlflow>=2.8.0,<2.10.0",  # Fixed: removed invalid [sqlalchemy] extra
    "scikit-learn>=1.5.2,<1.8.0",
    "pandas>=2.0.0,<2.2.0",
    "numpy>=1.26.0,<2.4.0",  # Ensure numpy >=1.26.0 for Python 3.12+ compatibility
    "scipy>=1.12.0,<1.17.0",
    # PyMC stack - Updated for better compatibility
    "pymc>=5.10.0,<5.17.0",  # Use newer PyMC version
    "pytensor>=2.18.0,<2.26.0",  # Explicitly constrain pytensor to compatible version
    "arviz>=0.17.0,<0.19.0",  # Updated arviz version
    "requests>=2.31.0,<2.33.0",
    "pyyaml>=6.0,<7.0",
    # JAX stack - Updated for compatibility
    "jax[cpu]>=0.4.20,<0.5.0",  # Use stable JAX version
    "jaxlib>=0.4.20,<0.5.0",    # Match jaxlib with jax
    "numpyro>=0.15.0,<0.16.0",  # Updated numpyro
    "psutil>=5.0.0,<8.0.0",
    # Additional dependencies for compatibility
    "packaging>=21.0",
    "toml>=0.10.2",
    # Add setuptools explicitly for Python 3.12+
    "setuptools>=68.0.0",  # Ensures setuptools availability
    "ipykernel>=6.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0,<8.0.0",
    "pytest-asyncio>=0.21.0,<0.23.0",
    "httpx>=0.24.0,<0.26.0"
]

[build-system]
requires = ["hatchling", "setuptools>=68.0.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# Environment management configuration
[tool.dependency-check]
python-version-range = ">=3.11,<3.13"
strict-mode = false
critical-packages = [
    "numpy", "scipy", "scikit-learn", "pandas", "mlflow", "pymc", "pytensor"
]

[tool.model-compatibility]
# Updated compatibility matrix
supported-sklearn-versions = ["1.5.2", "1.6.0", "1.7.0"]
supported-numpy-versions = ["1.26.0", "1.26.4", "2.0.0", "2.1.0", "2.2.0", "2.3.1"]
supported-scipy-versions = ["1.12.0", "1.13.0", "1.14.0", "1.15.0", "1.16.0"]
supported-pymc-versions = ["5.10.0", "5.11.0", "5.12.0", "5.13.0", "5.14.0", "5.15.0", "5.16.0"]
supported-pytensor-versions = ["2.18.0", "2.19.0", "2.20.0", "2.21.0", "2.22.0", "2.23.0", "2.24.0", "2.25.0"]
